requirements.txt
torch
torchvision
numpy
opencv-python
pillow
scikit-image
